Added YouTube functionality!
Users may now "-search" for YouTube search terms, and the bot will begin playing the first result it finds!

Functionality with "-join" and "-leave" has also been improved. If you have this bot "-join" your channel, it will not leave until the user types in "-leave", and will still be able to play sounds and music in the meantime.

Closes #3. Closes #4.